phantom_scraper/requirements.txt
selenium>=4.15.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pandas>=2.2.0
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.5.0
celery>=5.3.0
redis>=5.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
fake-useragent>=1.4.0
undetected-chromedriver>=3.5.0
webdriver-manager>=4.0.0
feedparser>=6.0.10
sqlalchemy>=2.0.0
cryptography>=41.0.0
